chore: Update Query view for side by side (#30655)
Updates view of Query / Api and JS to not show the header and right side
panels in side by side view

fixes: #30658
